requirements.txt
beautifulsoup4==4.11.1
browser-cookie3==0.14.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
flask==3.1.2
idna==3.3
importlib-metadata==4.11.3
jeepney==0.8.0
keyring==23.5.0
lz4==4.0.0
pbkdf2==1.3
pyaes==1.6.1
pycparser==2.21
pycryptodome==3.14.1
requests==2.27.1
SecretStorage==3.3.2
soupsieve==2.3.2.post1
tqdm==4.64.0
urllib3==1.26.9
zipp==3.8.0
python-dotenv==1.0.0
